% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerouac.R
\docType{package}
\name{kerouac}
\alias{kerouac}
\alias{kerouac-package}
\title{kerouac}
\usage{
kerouac()
}
\description{
Port in R/shiny of The Most Dangerous Writing App by Manuel Elbert.

Just try it with \code{kerouac()}
}
\references{
\code{http://www.themostdangerouswritingapp.com/}
}

